Fix to support 64-bit

+++ BlockAlertsDemo/ToAddToYourProjects/BlockActionSheet.m
@@ -249,7 +249,7 @@ static UIFont *buttonFont = nil;
 - (void)buttonClicked:(id)sender 
 {
     /* Run the button's block */
-    int buttonIndex = [sender tag] - 1;
+    NSInteger buttonIndex = [sender tag] - 1;
     [self dismissWithClickedButtonIndex:buttonIndex animated:YES];
 }
