+++ BlockAlertsDemo/ToAddToYourProjects/BlockAlertView.m
@@ -426,7 +426,7 @@ static UIFont *buttonFont = nil;
 - (void)buttonClicked:(id)sender 
 {
     /* Run the button's block */
-    int buttonIndex = [sender tag] - 1;
+    NSInteger buttonIndex = [sender tag] - 1;
     [self dismissWithClickedButtonIndex:buttonIndex animated:YES];
 }
